requirements.txt
asgiref==3.9.1
Django==5.2.5
sqlparse==0.5.3
gunicorn==21.2.0
psycopg2-binary==2.9.10
whitenoise==6.6.0
python-dotenv==1.0.0
dj-database-url==2.1.0
django-allauth==0.57.0
requests==2.32.5
python-decouple==3.8
